+++ dwds/CHANGELOG.md
@@ -13,6 +13,7 @@
 
 - Fix chrome detection in iPhone emulation mode in chrome or edge browsers.
 - Reliably find unused port for extension backend http service.
+- Ignore offset / count parameters in getObject if the object has no length
 
 ## 11.4.0